requirements.txt
arrow==1.2.3
binaryornot==0.4.4
briefcase==0.3.14
build==0.10.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==2.1.1
exceptiongroup==1.1.1
gbulb==0.6.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jinja2-time==0.2.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
packaging==23.1
platformdirs==3.5.0
pluggy==1.0.0
psutil==5.9.5
pycairo==1.23.0
Pygments==2.15.1
PyGObject==3.44.1
pyproject_hooks==1.0.0
pytest==7.3.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytube==12.1.3
PyYAML==6.0
requests==2.29.0
rich==13.3.5
six==1.16.0
smmap==5.0.0
text-unidecode==1.3
toga-core==0.3.1
toga-gtk==0.3.1
tomli==2.0.1
tomli_w==1.0.0
travertino==0.2.0
urllib3==1.26.15
